+++ docker-compose.yaml
@@ -1,6 +1,6 @@
 services:
   db:
-    container_name: postgres-db
+    container_name: postgres
     image: postgres:latest
     restart: unless-stopped
 
@@ -20,7 +20,7 @@ services:
       test: [ "CMD-SHELL", "pg_isready -U d17_map" ]
 
   web:
-    container_name: ktor-server
+    container_name: ktor
     build: .
     ports:
       - "8080:8080"
