+++ gradle.properties
@@ -1,7 +1,9 @@
 ktor_version=2.3.10
 kotlin_version=1.9.23
 logback_version=1.4.14
-kotlin.code.style=official
 
 postgresql_driver_version=42.7.3
-exposed_version = 0.41.1
+exposed_version=0.41.1
+
+h2_version=2.1.214
+kotlin.code.style=official
