+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "d17_map"
+rootProject.name = "d17_map-backend"
