+++ src/main/kotlin/inc/evil/plugins/Serialization.kt
@@ -1,4 +1,4 @@
-package com.example.plugins
+package inc.evil.plugins
 
 import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
