+++ src/main/resources/application.yaml
@@ -1,0 +1,6 @@
+ktor:
+    application:
+        modules:
+            - inc.evil.ApplicationKt.module
+    deployment:
+        port: 8080
