+++ src/test/kotlin/inc/evil/ApplicationTest.kt
@@ -1,12 +1,12 @@
-package com.example
+package inc.evil
 
-import com.example.plugins.*
+import inc.evil.plugins.configureRouting
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
 import io.ktor.server.testing.*
 import junit.framework.TestCase.assertEquals
-import kotlin.test.*
+import kotlin.test.Test
 
 class ApplicationTest {
     @Test
